chore: fix example
