+++ projects/addon-mobile/styles/ios/app-bar.less
@@ -1,6 +1,5 @@
 ._liquid-glass,
-&._liquid-glass,
-._liquid-glass & {
+&._liquid-glass {
     tui-app-bar {
         --t-padding: 1rem;
