+++ projects/demo/src/pages/components/app-bar/examples/1/index.html
@@ -1,109 +1,101 @@
 <h3>iOS</h3>
 
-@if (option !== 'stable' && option.liquidGlass?.()) {
-    <ng-container *ngTemplateOutlet="ios" />
-} @else {
-    <ng-container *ngTemplateOutlet="ios" />
-}
+<section tuiPlatform="ios">
+    <tui-app-bar>
+        <button
+            tuiButton
+            tuiSlot="start"
+            type="button"
+        >
+            Back
+        </button>
+        Taiga UI
+        <button
+            tuiButton
+            tuiSlot="end"
+            type="button"
+        >
+            Action
+        </button>
+    </tui-app-bar>
 
-<ng-template #ios>
-    <section tuiPlatform="ios">
-        <tui-app-bar>
-            <button
-                tuiButton
-                tuiSlot="start"
-                type="button"
-            >
-                Back
-            </button>
+    <tui-app-bar>
+        <button
+            tuiAppBarBack
+            tuiSlot="start"
+            type="button"
+        >
+            Back
+        </button>
+        Taiga UI — Components library
+        <button
+            iconStart="@tui.user"
+            tuiIconButton
+            tuiSlot="end"
+            type="button"
+        >
+            User
+        </button>
+    </tui-app-bar>
+
+    <tui-app-bar>
+        <button
+            title="Back"
+            tuiAppBarBack
+            tuiSlot="start"
+            type="button"
+        ></button>
+        Taiga UI
+        <button
+            iconStart="@tui.user"
+            tuiIconButton
+            tuiSlot="end"
+            type="button"
+        >
+            User
+        </button>
+        <button
+            iconStart="@tui.settings"
+            title="Settings"
+            tuiIconButton
+            tuiSlot="end"
+            type="button"
+        ></button>
+    </tui-app-bar>
+
+    <tui-app-bar>
+        <label tuiTitle>
             Taiga UI
-            <button
-                tuiButton
-                tuiSlot="end"
-                type="button"
-            >
-                Action
-            </button>
-        </tui-app-bar>
+            <span tuiSubtitle>Components library</span>
+        </label>
+        <button
+            iconStart="@tui.settings"
+            title="Settings"
+            tuiIconButton
+            tuiSlot="end"
+            type="button"
+        ></button>
+    </tui-app-bar>
 
-        <tui-app-bar>
-            <button
-                tuiAppBarBack
-                tuiSlot="start"
-                type="button"
-            >
-                Back
-            </button>
-            Taiga UI — Components library
-            <button
-                iconStart="@tui.user"
-                tuiIconButton
-                tuiSlot="end"
-                type="button"
-            >
-                User
-            </button>
-        </tui-app-bar>
+    <tui-app-bar>
+        <button
+            tuiAppBarBack
+            tuiSlot="start"
+            type="button"
+        >
+            Back
+        </button>
 
-        <tui-app-bar>
-            <button
-                title="Back"
-                tuiAppBarBack
-                tuiSlot="start"
-                type="button"
-            ></button>
-            Taiga UI
-            <button
-                iconStart="@tui.user"
-                tuiIconButton
-                tuiSlot="end"
-                type="button"
-            >
-                User
-            </button>
-            <button
-                iconStart="@tui.settings"
-                title="Settings"
-                tuiIconButton
-                tuiSlot="end"
-                type="button"
-            ></button>
-        </tui-app-bar>
-
-        <tui-app-bar>
-            <label tuiTitle>
-                Taiga UI
-                <span tuiSubtitle>Components library</span>
-            </label>
-            <button
-                iconStart="@tui.settings"
-                title="Settings"
-                tuiIconButton
-                tuiSlot="end"
-                type="button"
-            ></button>
-        </tui-app-bar>
-
-        <tui-app-bar>
-            <button
-                tuiAppBarBack
-                tuiSlot="start"
-                type="button"
-            >
-                Back
-            </button>
-
-            <div>
-                <progress
-                    size="s"
-                    tuiProgressBar
-                    [max]="100"
-                    [value]="35"
-                ></progress>
-            </div>
-        </tui-app-bar>
-    </section>
-</ng-template>
+        <div>
+            <progress
+                size="s"
+                tuiProgressBar
+                [max]="100"
+                [value]="35"
+            ></progress>
+        </div>
+    </tui-app-bar>
+</section>
 
 <h3 class="tui-space_top-8">Android</h3>
 <section tuiPlatform="android">
