+++ projects/demo/src/pages/components/app-bar/examples/1/index.less
@@ -3,7 +3,7 @@ tui-app-bar {
     margin-block-end: 1rem;
     box-shadow: var(--tui-shadow-small);
 
-    :host-context(._liquid-glass) & {
+    :host-context([data-platform='ios']._liquid-glass) & {
         box-shadow: none;
     }
 }
