+++ projects/demo/src/pages/components/app-bar/examples/1/index.ts
@@ -1,26 +1,16 @@
-import {Component, inject} from '@angular/core';
+import {Component} from '@angular/core';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
 import {TuiPlatform} from '@taiga-ui/cdk';
-import {TUI_OPTIONS, TuiButton, TuiTitle} from '@taiga-ui/core';
+import {TuiButton, TuiTitle} from '@taiga-ui/core';
 import {TuiProgressBar} from '@taiga-ui/kit';
 import {TuiAppBar} from '@taiga-ui/layout';
-import {NgTemplateOutlet} from '@angular/common';
 
 @Component({
-    imports: [
-        TuiAppBar,
-        TuiButton,
-        TuiPlatform,
-        TuiProgressBar,
-        TuiTitle,
-        NgTemplateOutlet,
-    ],
+    imports: [TuiAppBar, TuiButton, TuiPlatform, TuiProgressBar, TuiTitle],
     templateUrl: './index.html',
     styleUrl: './index.less',
     encapsulation,
     changeDetection,
 })
-export default class Example {
-    protected readonly option = inject(TUI_OPTIONS).apis;
-}
+export default class Example {}
