+++ projects/demo/src/pages/components/app-bar/examples/5/component.less
@@ -1,0 +1,17 @@
+tui-app-bar {
+    position: sticky;
+    z-index: 1;
+    inset-block-start: -0.03125rem;
+    margin-block-end: 0.75rem;
+    background: var(--tui-background-base);
+}
+
+:host-context([data-platform='ios']._liquid-glass) {
+    [tuiHeader] {
+        transition: opacity 300ms linear;
+    }
+
+    [tuiHeader]._hidden {
+        opacity: 0;
+    }
+}
