+++ projects/demo/src/pages/components/app-bar/examples/5/index.html
@@ -18,7 +18,7 @@
     tuiDynamicHeaderContainer
     waIntersectionRootMargin="-120px 0px 1000000% 0px"
     class="container liquid-glass"
-    [class._liquid-glass_blur]="container.scrollTop > 2"
+    [class._liquid-glass_blur]="container.scrollTop > 0"
     [tuiPlatform]="'ios'"
 >
     <ng-container *polymorpheusOutlet="content" />
