+++ projects/demo/src/pages/components/app-bar/examples/5/index.less
@@ -18,7 +18,7 @@ tui-segmented {
     display: none;
 }
 
-:host-context(._liquid-glass) {
+:host-context([data-platform='ios']._liquid-glass) {
     .default {
         display: none;
     }
