+++ projects/demo/src/pages/components/app-bar/examples/7/index.less
@@ -1,0 +1,35 @@
+.container {
+    max-block-size: 30rem;
+    inline-size: 20rem;
+    overflow: scroll;
+    overscroll-behavior: none;
+}
+
+tui-app-bar {
+    position: sticky;
+    z-index: 1;
+    inset-block-start: -0.03125rem;
+    margin-block-end: 0.75rem;
+}
+
+.content {
+    padding: 0;
+
+    > [tuiHeader] {
+        padding: 1rem;
+    }
+}
+
+.block {
+    block-size: 20rem;
+}
+
+[data-platform='ios'] {
+    .content > [tuiHeader] {
+        transition: opacity 300ms linear;
+    }
+
+    .content > [tuiHeader]._hidden {
+        opacity: 0;
+    }
+}
