+++ projects/demo/src/pages/components/app-bar/examples/7/index.ts
@@ -1,0 +1,35 @@
+import {Component, inject, signal} from '@angular/core';
+import {changeDetection} from '@demo/emulate/change-detection';
+import {encapsulation} from '@demo/emulate/encapsulation';
+import {WaIntersectionObserver} from '@ng-web-apis/intersection-observer';
+import {TuiPlatform} from '@taiga-ui/cdk';
+import {TUI_DARK_MODE, TuiTitle} from '@taiga-ui/core';
+import {TuiAppBar, TuiDynamicHeader, TuiHeader} from '@taiga-ui/layout';
+
+@Component({
+    imports: [
+        TuiAppBar,
+        TuiDynamicHeader,
+        TuiHeader,
+        TuiPlatform,
+        TuiTitle,
+        WaIntersectionObserver,
+    ],
+    templateUrl: './index.html',
+    styleUrl: './index.less',
+    encapsulation,
+    changeDetection,
+})
+export default class Example {
+    protected readonly mode = inject(TUI_DARK_MODE);
+    protected readonly colors = ['#2f3f5b', '#3a3f66', '#5a2f3f', '#3f4e37', '#5a3851'];
+    protected readonly color = signal('');
+
+    protected onIntersection(isIntersecting: boolean, color: string): void {
+        if (isIntersecting) {
+            this.color.set(color);
+        } else if (this.color() === color) {
+            this.color.set('');
+        }
+    }
+}
