+++ projects/layout/components/app-bar/app-bar-back.style.less
@@ -35,7 +35,7 @@
             display: none;
         }
 
-        :host-context(._liquid-glass) & {
+        :host-context([data-platform='ios']._liquid-glass) & {
             display: flex;
             font-size: 0;
             block-size: 2.75rem;
