+++ projects/layout/components/app-bar/liquid-glass/app-bar-button.style.less
@@ -1,4 +1,4 @@
-._liquid-glass [tuiAppBarButton]:where(*&) {
+[data-platform='ios']._liquid-glass [tuiAppBarButton]:where(*&) {
     border-radius: 2.75rem;
     margin-inline: 1rem;
     backdrop-filter: blur(0.25rem) saturate(6) brightness(1.05);
@@ -54,8 +54,8 @@
     }
 }
 
-._liquid-glass [tuiTheme='dark'] [tuiAppBarButton]:where(*&),
-[tuiTheme='dark'] ._liquid-glass [tuiAppBarButton]:where(*&),
+[data-platform='ios']._liquid-glass [tuiTheme='dark'] [tuiAppBarButton]:where(*&),
+[tuiTheme='dark'] [data-platform='ios']._liquid-glass [tuiAppBarButton]:where(*&),
 [tuiTheme='dark']._liquid-glass [tuiAppBarButton]:where(*&) {
     --t-appbar-button-bg: rgba(88, 88, 88, 0.31);
     --t-appbar-glow: var(--tui-appbar-background, rgba(31, 31, 31, 0.8));
@@ -68,6 +68,6 @@
     }
 }
 
-._liquid-glass tui-app-bar[tuiTheme='dark'] [tuiAppBarButton]:where(*&) {
+[data-platform='ios']._liquid-glass tui-app-bar[tuiTheme='dark'] [tuiAppBarButton]:where(*&) {
     --t-appbar-bevel: linear-gradient(319deg, #fff 4%, #fff 19%, rgba(255, 255, 255, 0.12) 40%, rgba(255, 255, 255, 0.02) 45%, rgba(255, 255, 255, 0) 48%, rgba(255, 255, 255, 0) 50%, rgba(255, 255, 255, 0) 52%, rgba(255, 255, 255, 0.02) 55%, rgba(255, 255, 255, 0.12) 60%, #fff 80%, #fff 96%);
 }
